requirements.txt
cfscrape
urllib3==1.26.16
cloudscraper
httpx
pysocks
termcolor
requests
colorama

